+++ docs/.vitepress/theme/index.ts
@@ -1,4 +1,7 @@
+import type { Theme } from 'vitepress'
 import DefaultTheme from 'vitepress/theme'
-import './style.css'
+import './custom.css'
 
-export default DefaultTheme
+export default {
+  extends: DefaultTheme
+} satisfies Theme